pyproject.toml
[tool.poetry]
name = "sprinkler"
version = "0.1.0"
description = ""
authors = ["Patrick <pdassier@free.fr>"]

[tool.poetry.dependencies]
pubnub = "^7.0.1"
python-dateutil = "^2.8.2"
python = "^3.10"
jsonschema = "^4.17.0"
aiohttp = "^3.8.3"
env-file = "^2020.12.3"

[tool.poetry.dev-dependencies]
pylint = "^2.15.6"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
autopep8 = "^2.0.0"
pytest-asyncio = "^0.20.2"
ipython = "^8.6.0"
python-dotenv = "^0.21.0"
